+++ medialibrary/src/org/videolan/medialibrary/media/Album.java
@@ -11,7 +11,7 @@ import org.videolan.medialibrary.interfaces.media.AbstractMediaWrapper;
 public class Album extends AbstractAlbum {
     public static final String TAG = "VLC/Album";
 
-    public Album(long id, String title, int releaseYear, String artworkMrl, String albumArtist, long albumArtistId, int nbTracks, int duration) {
+    public Album(long id, String title, int releaseYear, String artworkMrl, String albumArtist, long albumArtistId, int nbTracks, long duration) {
         super(id, title, releaseYear, artworkMrl, albumArtist, albumArtistId, nbTracks, duration);
     }
